refactoring: moved Project tests to separate file (#16)
