+++ tests/project_test.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+import os
+import unittest
+from blameandshame.base import  Project
+
+
+class ProjectTestCase(unittest.TestCase):
+    def test_url_to_path(self):
+        repos_dir = os.path.join(os.getcwd(), '.repos')
+
+        self.assertEqual(Project._url_to_path('https://github.com/uber/pyro'),
+                         os.path.join(repos_dir, 'pyro'))
+        self.assertEqual(Project._url_to_path('https://github.com/google/protobuf'),
+                         os.path.join(repos_dir, 'protobuf'))
+        self.assertEqual(Project._url_to_path('https://github.com/cilium/cilium'),
+                         os.path.join(repos_dir, 'cilium'))
+        self.assertEqual(Project._url_to_path('https://github.com/golang/dep'),
+                         os.path.join(repos_dir, 'dep'))
+        self.assertEqual(Project._url_to_path('https://github.com/opencv/opencv.git'),
+                         os.path.join(repos_dir, 'opencv'))
+
+
+if __name__ == '__main__':
+    unittest.main()
